backend/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
flask-cors==4.0.0
twilio==8.12.0
PyJWT==2.8.0
pytest==7.4.3
requests==2.31.0
flask-jwt-extended==4.7.1
plaid-python==16.0.0
flask-pymongo==2.3.0
pymongo==4.6.1